OBVGGT/demo_a/requirements.txt
gradio>=4.0
torch
torchvision
numpy
matplotlib
plotly
Pillow
scipy
opencv-python
trimesh
huggingface_hub
